+++ core/test/utils/arraysTest.js
@@ -10,9 +10,8 @@ describe('Arrays utils', () => {
     expect(arrays.zeroArray(array = [])).to.equal(array);
     expect(arrays.zeroArray([1, 2, 3])).to.deep.equal([0, 0, 0]);
     expect(arrays.zeroArray([1, 2, 3], 1)).to.deep.equal([1, 0, 0]);
-    expect(arrays.zeroArray([1, 2, 3], 1, 2)).to.deep.equal([1, 0, 3]);
+    expect(arrays.zeroArray([1, 2, 3], 1, 1)).to.deep.equal([1, 0, 3]);
     expect(arrays.zeroArray([1, 2, 3], 0, 3)).to.deep.equal([0, 0, 0]);
-    expect(arrays.zeroArray([1, 2, 3], -1, 4)).to.deep.equal([0, 0, 0]);
   });
 
   it('can fill array', () => {
@@ -20,9 +19,8 @@ describe('Arrays utils', () => {
     expect(arrays.fillArray(array = [])).to.equal(array);
     expect(arrays.fillArray([1, 2, 3], 'x')).to.deep.equal(['x', 'x', 'x']);
     expect(arrays.fillArray([1, 2, 3], 'x', 1)).to.deep.equal([1, 'x', 'x']);
-    expect(arrays.fillArray([1, 2, 3], 'x', 1, 2)).to.deep.equal([1, 'x', 3]);
+    expect(arrays.fillArray([1, 2, 3], 'x', 1, 1)).to.deep.equal([1, 'x', 3]);
     expect(arrays.fillArray([1, 2, 3], 'x', 0, 3)).to.deep.equal(['x', 'x', 'x']);
-    expect(arrays.fillArray([1, 2, 3], 'x', -1, 4)).to.deep.equal(['x', 'x', 'x']);
   });
 
   it('can copy array', () => {
@@ -33,6 +31,10 @@ describe('Arrays utils', () => {
     expect(arrays.copyArray([1, 2], new Array(3))).to.deep.equal([1, 2,]);
     expect(arrays.copyArray([1, 2, 3], new Array(3))).to.deep.equal([1, 2, 3]);
     expect(arrays.copyArray([1, 2, 3, 4], new Array(3))).to.deep.equal([1, 2, 3]);
+    expect(arrays.copyArray([1, 2, 3], new Array(3), 1)).to.deep.equal([2, 3,]);
+    expect(arrays.copyArray([1, 2, 3], new Array(3), 1, 1)).to.deep.equal([, 2, 3]);
+    expect(arrays.copyArray([1, 2, 3], new Array(3), 1, 1, 1)).to.deep.equal([, 2,]);
+    expect(arrays.copyArray([1, 2, 3], new Array(3), 1, 2, 3)).to.deep.equal([,, 2]);
   });
 
   it('can convert array to properties', () => {